Split `qpd.py` code into three files
