+++ circuit_knitting/cutting/qpd/__init__.py
@@ -12,12 +12,12 @@
 """Main quasiprobability decomposition functionality."""
 
 from .qpd_basis import QPDBasis
-from .qpd import (
-    generate_qpd_weights,
-    decompose_qpd_instructions,
+from .weights import (
     WeightType,
-    qpdbasis_from_instruction,
+    generate_qpd_weights,
 )
+from .decompose import decompose_qpd_instructions
+from .decompositions import qpdbasis_from_instruction
 from .instructions import (
     BaseQPDGate,
     SingleQubitQPDGate,
